fix(resources): makes `kv` property optional for `DenoKvResource` (closes #124)
Making the `kv` instance a required property for `DenoKvResource` options was causing issues with `Deno` leaking to the browser. After refactoring to try out different project structures, it was hard to avoid leaks, which result in a very poor UX (and DX). Making the `kv` property optional not only fixes these issue altogether, it also avoids having to explicitly instantiate a `kv` instance via `Deno.openKv()` in another file e.g. `db.ts`. Noteworthy is, that in the context of netzo, interacting with `kv` directly is actually not desired, as all interactions should be made via resources and/or the API itself to benefit from resource methods, hooks, resolvers, etc. So this closes #124 for now.